requirements.txt
asgiref==3.8.1
Django==5.1.6
image==1.5.33
mpmath==1.3.0
pillow==11.1.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
